# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AdminUserProfileStatus = typeof AdminUserProfileStatus[keyof typeof AdminUserProfileStatus];


export const AdminUserProfileStatus = {
  Active: 'Active',
  Suspended: 'Suspended',
} as const;

export interface AdminUserProfile {
  userId: string;
  status: AdminUserProfileStatus;
  notes: string;
  updatedAt: string;
}

export interface AdminUserNotesUpdate {
  /** @maxLength 5000 */
  notes: string;
}

export type AdminUserStatusUpdateStatus = typeof AdminUserStatusUpdateStatus[keyof typeof AdminUserStatusUpdateStatus];


export const AdminUserStatusUpdateStatus = {
  Active: 'Active',
  Suspended: 'Suspended',
} as const;

export interface AdminUserStatusUpdate {
  status: AdminUserStatusUpdateStatus;
}

export interface AdminUserActionResponse {
  success: boolean;
  message: string;
}

export interface ApiError {
  title: string;
  detail: string;
}

export type InvestmentPlanInputStatus = typeof InvestmentPlanInputStatus[keyof typeof InvestmentPlanInputStatus];


export const InvestmentPlanInputStatus = {
  Active: 'Active',
  Disabled: 'Disabled',
} as const;

export interface InvestmentPlanInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  minInvestment: number;
  /**
     * @minimum 0
     * @nullable
     */
  maxInvestment: number | null;
  /** @minimum 0 */
  profitPercentage: number;
  /** @minLength 1 */
  executionCycle: string;
  description: string;
  /** @minItems 1 */
  features: string[];
  status: InvestmentPlanInputStatus;
  /** @minimum 0 */
  displayOrder: number;
}

export type InvestmentPlan = InvestmentPlanInput & {
  id: string;
  /** @minimum 0 */
  investors: number;
  /** @minimum 0 */
  totalDeposited: number;
};

export interface InvestmentUser {
  id: string;
  name: string;
  email: string;
}

export interface InvestmentPlanSnapshot {
  id: string;
  name: string;
  executionCycle: string;
}

export type InvestmentStatus = typeof InvestmentStatus[keyof typeof InvestmentStatus];


export const InvestmentStatus = {
  Pending: 'Pending',
  Active: 'Active',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export type InvestmentDisplayStatus = typeof InvestmentDisplayStatus[keyof typeof InvestmentDisplayStatus];


export const InvestmentDisplayStatus = {
  Pending: 'Pending',
  Active: 'Active',
  Paused: 'Paused',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface Investment {
  id: string;
  user: InvestmentUser;
  plan: InvestmentPlanSnapshot;
  /** @minimum 0 */
  investmentAmount: number;
  /** @minimum 0 */
  profitPercentage: number;
  investmentDate: string;
  maturityDate: string;
  status: InvestmentStatus;
  isPaused: boolean;
  /** @minimum 0 */
  expectedReturn: number;
  /** @minimum 0 */
  remainingSeconds: number;
  displayStatus: InvestmentDisplayStatus;
}

export interface InvestmentInput {
  /** @minLength 1 */
  planId: string;
  /** @exclusiveMinimum 0 */
  amount: number;
}

export type InvestmentStatusActionAction = typeof InvestmentStatusActionAction[keyof typeof InvestmentStatusActionAction];


export const InvestmentStatusActionAction = {
  activate: 'activate',
  pause: 'pause',
  complete: 'complete',
  cancel: 'cancel',
} as const;

export interface InvestmentStatusAction {
  action: InvestmentStatusActionAction;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ApiError;

/**
 * Resource not found
 */
export type NotFoundResponse = ApiError;

/**
 * Request conflicts with the current resource state
 */
export type ConflictResponse = ApiError;

export type ListInvestmentsParams = {
scope?: ListInvestmentsScope;
userId?: string;
};

export type ListInvestmentsScope = typeof ListInvestmentsScope[keyof typeof ListInvestmentsScope];


export const ListInvestmentsScope = {
  user: 'user',
  all: 'all',
} as const;