# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin-users
    description: Admin user profile and account management operations
  - name: plans
    description: Investment plan catalog management operations
  - name: investments
    description: User investment assignment and lifecycle operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /admin/users/{userId}/profile:
    get:
      operationId: getAdminUserProfile
      tags: [admin-users]
      summary: Get an admin user profile
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Profile state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProfile"
  /admin/users/{userId}/notes:
    put:
      operationId: updateAdminUserNotes
      tags: [admin-users]
      summary: Save private admin notes
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserNotesUpdate"
      responses:
        "200":
          description: Updated profile state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
  /admin/users/{userId}/status:
    patch:
      operationId: updateAdminUserStatus
      tags: [admin-users]
      summary: Activate or suspend a user account
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserStatusUpdate"
      responses:
        "200":
          description: Updated profile state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
  /admin/users/{userId}/password-reset:
    post:
      operationId: requestAdminUserPasswordReset
      tags: [admin-users]
      summary: Request a user password reset
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserActionResponse"
  /admin/users/{userId}:
    delete:
      operationId: deleteAdminUser
      tags: [admin-users]
      summary: Delete a user account
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Delete request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserActionResponse"
  /plans:
    get:
      operationId: listInvestmentPlans
      tags: [plans]
      summary: List investment plans
      responses:
        "200":
          description: Investment plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestmentPlan"
    post:
      operationId: createInvestmentPlan
      tags: [plans]
      summary: Create an investment plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentPlanInput"
      responses:
        "201":
          description: Created investment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
        "400":
          $ref: "#/components/responses/BadRequest"
  /plans/{planId}:
    get:
      operationId: getInvestmentPlan
      tags: [plans]
      summary: Get an investment plan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Investment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateInvestmentPlan
      tags: [plans]
      summary: Update an investment plan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentPlanInput"
      responses:
        "200":
          description: Updated investment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteInvestmentPlan
      tags: [plans]
      summary: Delete an investment plan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Delete request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserActionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /plans/{planId}/status:
    patch:
      operationId: updateInvestmentPlanStatus
      tags: [plans]
      summary: Enable or disable an investment plan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserStatusUpdate"
      responses:
        "200":
          description: Updated investment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentPlan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List investments
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [user, all]
            default: user
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Investment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a pending investment assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Created investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /investments/{investmentId}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get an investment
      parameters:
        - $ref: "#/components/parameters/InvestmentId"
      responses:
        "200":
          description: Investment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "404":
          $ref: "#/components/responses/NotFound"
  /investments/{investmentId}/status:
    patch:
      operationId: updateInvestmentStatus
      tags: [investments]
      summary: Apply an admin lifecycle action to an investment
      parameters:
        - $ref: "#/components/parameters/InvestmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentStatusAction"
      responses:
        "200":
          description: Updated investment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: Application user identifier, for example #U-4821
      schema:
        type: string
        minLength: 1
    PlanId:
      name: planId
      in: path
      required: true
      description: Investment plan identifier
      schema:
        type: string
        minLength: 1
    InvestmentId:
      name: investmentId
      in: path
      required: true
      description: Investment record identifier
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Request conflicts with the current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AdminUserProfile:
      type: object
      required: [userId, status, notes, updatedAt]
      properties:
        userId:
          type: string
        status:
          type: string
          enum: [Active, Suspended]
        notes:
          type: string
        updatedAt:
          type: string
          format: date-time
    AdminUserNotesUpdate:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
          maxLength: 5000
    AdminUserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Active, Suspended]
    AdminUserActionResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    ApiError:
      type: object
      required: [title, detail]
      properties:
        title:
          type: string
        detail:
          type: string
    InvestmentPlan:
      allOf:
        - $ref: "#/components/schemas/InvestmentPlanInput"
        - type: object
          required: [id, investors, totalDeposited]
          properties:
            id:
              type: string
            investors:
              type: integer
              minimum: 0
            totalDeposited:
              type: number
              minimum: 0
    InvestmentPlanInput:
      type: object
      required: [name, minInvestment, maxInvestment, profitPercentage, executionCycle, description, features, status, displayOrder]
      properties:
        name:
          type: string
          minLength: 1
        minInvestment:
          type: number
          minimum: 0
        maxInvestment:
          type:
            - number
            - "null"
          minimum: 0
        profitPercentage:
          type: number
          minimum: 0
        executionCycle:
          type: string
          minLength: 1
        description:
          type: string
        features:
          type: array
          items:
            type: string
          minItems: 1
        status:
          type: string
          enum: [Active, Disabled]
        displayOrder:
          type: integer
          minimum: 0
    InvestmentUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    InvestmentPlanSnapshot:
      type: object
      required: [id, name, executionCycle]
      properties:
        id:
          type: string
        name:
          type: string
        executionCycle:
          type: string
    Investment:
      type: object
      required: [id, user, plan, investmentAmount, profitPercentage, investmentDate, maturityDate, status, isPaused, expectedReturn, remainingSeconds, displayStatus]
      properties:
        id:
          type: string
        user:
          $ref: "#/components/schemas/InvestmentUser"
        plan:
          $ref: "#/components/schemas/InvestmentPlanSnapshot"
        investmentAmount:
          type: number
          minimum: 0
        profitPercentage:
          type: number
          minimum: 0
        investmentDate:
          type: string
          format: date-time
        maturityDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [Pending, Active, Completed, Cancelled]
        isPaused:
          type: boolean
        expectedReturn:
          type: number
          minimum: 0
        remainingSeconds:
          type: integer
          minimum: 0
        displayStatus:
          type: string
          enum: [Pending, Active, Paused, Completed, Cancelled]
    InvestmentInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: string
          minLength: 1
        amount:
          type: number
          exclusiveMinimum: 0
    InvestmentStatusAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [activate, pause, complete, cancel]