# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentDisplayStatus } from './investmentDisplayStatus';
import type { InvestmentPlanSnapshot } from './investmentPlanSnapshot';
import type { InvestmentStatus } from './investmentStatus';
import type { InvestmentUser } from './investmentUser';

export interface Investment {
  id: string;
  user: InvestmentUser;
  plan: InvestmentPlanSnapshot;
  /** @minimum 0 */
  investmentAmount: number;
  /** @minimum 0 */
  profitPercentage: number;
  investmentDate: Date;
  maturityDate: Date;
  status: InvestmentStatus;
  isPaused: boolean;
  /** @minimum 0 */
  expectedReturn: number;
  /** @minimum 0 */
  remainingSeconds: number;
  displayStatus: InvestmentDisplayStatus;
}