# lib/api-zod/src/generated/types/investmentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvestmentInput {
  /** @minLength 1 */
  planId: string;
  /** @exclusiveMinimum 0 */
  amount: number;
}