# lib/api-zod/src/generated/types/investmentPlanSnapshot.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvestmentPlanSnapshot {
  id: string;
  name: string;
  executionCycle: string;
}