# lib/api-zod/src/generated/types/investmentStatusAction.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestmentStatusActionAction } from './investmentStatusActionAction';

export interface InvestmentStatusAction {
  action: InvestmentStatusActionAction;
}